pyproject.toml
[project]
name = "gc-doc-mcp"
version = "1.0.0"
description = "GrapeCity Documentation MCP Service - RAG-based document search for SpreadJS and GcExcel"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
	{name = "GrapeCity"}
]
keywords = ["mcp", "rag", "vector-search", "documentation", "spreadjs", "gcexcel"]

dependencies = [
	"voyageai>=0.3.0",
	"qdrant-client>=1.9.0",
	"fastembed>=0.3.0",
	"lingua-language-detector>=2.0.0",
	"fastapi>=0.110.0",
	"uvicorn>=0.27.0",
	"pydantic>=2.0.0",
	"pydantic-settings>=2.0.0",
	"python-multipart>=0.0.9",
	"httpx>=0.27.0",
	"pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
	"pytest>=8.0.0",
	"pytest-asyncio>=0.23.0",
	"black>=24.0.0",
	"ruff>=0.3.0",
	"mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
